requirements-dev.txt
coverage==5.0.4
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
nose==1.3.7
pycodestyle==2.5.0
pyflakes==2.1.1
PyMySQL==0.9.3
SQLAlchemy==1.3.15
testing.common.database==2.0.3
testing.mysqld==1.4.0
